Run unit tests using built wheel
