+++ tests/unit_tests/prompts/test_few_shot.py
@@ -1,5 +1,4 @@
 """Test few shot prompt template."""
-from typing import Dict, List, Tuple
 
 import pytest
 
@@ -9,25 +8,6 @@ from langchain.prompts.prompt import PromptTemplate
 EXAMPLE_PROMPT = PromptTemplate(
     input_variables=["question", "answer"], template="{question}: {answer}"
 )
-
-
-@pytest.fixture()
-def example_jinja2_prompt() -> Tuple[PromptTemplate, List[Dict[str, str]]]:
-    example_template = "{{ word }}: {{ antonym }}"
-
-    examples = [
-        {"word": "happy", "antonym": "sad"},
-        {"word": "tall", "antonym": "short"},
-    ]
-
-    return (
-        PromptTemplate(
-            input_variables=["word", "antonym"],
-            template=example_template,
-            template_format="jinja2",
-        ),
-        examples,
-    )
 
 
 def test_suffix_only() -> None:
@@ -195,71 +175,3 @@ def test_partial() -> None:
         "Now you try to talk about party."
     )
     assert output == expected_output
-
-
-def test_prompt_jinja2_functionality(
-    example_jinja2_prompt: Tuple[PromptTemplate, List[Dict[str, str]]]
-) -> None:
-    prefix = "Starting with {{ foo }}"
-    suffix = "Ending with {{ bar }}"
-
-    prompt = FewShotPromptTemplate(
-        input_variables=["foo", "bar"],
-        suffix=suffix,
-        prefix=prefix,
-        examples=example_jinja2_prompt[1],
-        example_prompt=example_jinja2_prompt[0],
-        template_format="jinja2",
-    )
-    output = prompt.format(foo="hello", bar="bye")
-    expected_output = (
-        "Starting with hello\n\n" "happy: sad\n\n" "tall: short\n\n" "Ending with bye"
-    )
-
-    assert output == expected_output
-
-
-def test_prompt_jinja2_missing_input_variables(
-    example_jinja2_prompt: Tuple[PromptTemplate, List[Dict[str, str]]]
-) -> None:
-    """Test error is raised when input variables are not provided."""
-    prefix = "Starting with {{ foo }}"
-    suffix = "Ending with {{ bar }}"
-
-    # Test when missing in suffix
-    with pytest.raises(ValueError):
-        FewShotPromptTemplate(
-            input_variables=[],
-            suffix=suffix,
-            examples=example_jinja2_prompt[1],
-            example_prompt=example_jinja2_prompt[0],
-            template_format="jinja2",
-        )
-
-    # Test when missing in prefix
-    with pytest.raises(ValueError):
-        FewShotPromptTemplate(
-            input_variables=["bar"],
-            suffix=suffix,
-            prefix=prefix,
-            examples=example_jinja2_prompt[1],
-            example_prompt=example_jinja2_prompt[0],
-            template_format="jinja2",
-        )
-
-
-def test_prompt_jinja2_extra_input_variables(
-    example_jinja2_prompt: Tuple[PromptTemplate, List[Dict[str, str]]]
-) -> None:
-    """Test error is raised when there are too many input variables."""
-    prefix = "Starting with {{ foo }}"
-    suffix = "Ending with {{ bar }}"
-    with pytest.raises(ValueError):
-        FewShotPromptTemplate(
-            input_variables=["bar", "foo", "extra", "thing"],
-            suffix=suffix,
-            prefix=prefix,
-            examples=example_jinja2_prompt[1],
-            example_prompt=example_jinja2_prompt[0],
-            template_format="jinja2",
-        )
